requirements.txt
fastapi
uvicorn
psutil
faiss-cpu
sentence-transformers
requests
streamlit